refactor(git-pull-request) coding conventions
