+++ packages/amplication-git-pull-request-service/.env
@@ -1,9 +1,7 @@
 # Kafka
 KAFKA_BROKERS='localhost:9092'
-GENERATE_PULL_REQUEST_TOPIC="git.internal.pull-request.request.1"
 KAFKA_CLIENT_ID="git-pull-request"
 KAFKA_GROUP_ID="amplication-git-pull-request-service"
-KAFKA_REPOSITORY_PUSH_QUEUE="git.external.push.event.0"
 
 # GitHub:
 GITHUB_APP_APP_ID=[github-app-app-id]
