+++ packages/amplication-git-pull-request-service/src/app.module.ts
@@ -1,18 +1,19 @@
 import { AmplicationLoggerModule } from '@amplication/nest-logger-module';
 import { Module, OnApplicationShutdown } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
-import { SERVICE_NAME } from './constants';
-import { CoreModule } from './core';
+import { DiffModule } from './diff/diff.module';
+import { PullRequestModule } from './pull-request/pull-request.module';
 
 @Module({
   imports: [
-    CoreModule,
+    DiffModule,
+    PullRequestModule,
     ConfigModule.forRoot({
       isGlobal: true,
       envFilePath: ['.env.local', '.env'],
     }),
     AmplicationLoggerModule.register({
-      metadata: { service: SERVICE_NAME },
+      metadata: { service: "amplication-git-pull-request-service" },
     }),
   ],
 })
