--- packages/amplication-git-pull-request-service/src/core/core.module.ts
@@ -1,7 +1,0 @@
-import { Module } from '@nestjs/common';
-import { PullRequestModule } from './pull-request';
-@Module({
-  imports: [PullRequestModule],
-  exports: [PullRequestModule],
-})
-export class CoreModule {}
