--- packages/amplication-git-pull-request-service/src/core/diff/index.ts
@@ -1,2 +1,0 @@
-export * from './diff.module';
-export * from './diff.service';
