--- packages/amplication-git-pull-request-service/src/core/diff/utils/BuildPathFactory.mock.ts
@@ -1,4 +1,0 @@
-import { mock } from 'jest-mock-extended';
-import { BuildPathFactory } from './BuildPathFactory';
-
-export const MOCK_BUILD_PATH_FACTORY = mock<BuildPathFactory>();
