--- packages/amplication-git-pull-request-service/src/core/index.ts
@@ -1,3 +1,0 @@
-export * from './diff';
-export * from './pull-request';
-export * from './core.module';
