--- packages/amplication-git-pull-request-service/src/core/pull-request/dto/StatusEnum.ts
@@ -1,4 +1,0 @@
-export enum StatusEnum {
-  Success,
-  GeneralFail,
-}
