--- packages/amplication-git-pull-request-service/src/core/pull-request/index.ts
@@ -1,1 +1,0 @@
-export * from './pull-request.module';
