+++ packages/amplication-git-pull-request-service/src/diff/diff.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { DiffService } from './diff.service';
-import { BuildPathFactory } from './utils/BuildPathFactory';
+import { BuildPathFactory } from './build-path-factory';
 
 @Module({
   providers: [DiffService, BuildPathFactory],
