+++ packages/amplication-git-pull-request-service/src/diff/diffset-mapper.ts
@@ -1,6 +1,6 @@
 import { DiffSet } from 'dir-compare';
-import { PrModule } from '../../constants';
-import { DiffVisitorFn } from './visitors/types';
+import { PrModule } from '../types';
+import { DiffVisitorFn } from './types';
 
 /**
  * This method takes the diffSets and visitor functions and returns an array of modules.
