+++ packages/amplication-git-pull-request-service/src/diff/types.ts
@@ -1,4 +1,4 @@
 import { Difference } from 'dir-compare';
-import { PrModule } from '../../../constants';
+import { PrModule } from '../types';
 
 export type DiffVisitorFn = (diff: Difference) => PrModule | null;
