+++ packages/amplication-git-pull-request-service/src/pull-request/pull-request.type.ts
@@ -1,0 +1,8 @@
+export enum KafkaTopics {
+  GeneralPullRequest = "git.internal.pull-request.request.1",
+}
+
+export enum StatusEnum {
+  Success,
+  GeneralFail,
+}
