+++ packages/amplication-git-pull-request-service/src/types.ts
@@ -1,0 +1,4 @@
+export type PrModule = {
+  path: string;
+  code: string | null;
+};
