+++ packages/amplication-git-pull-request-service/tests/diff/diffset-mapper.spec.ts
@@ -1,6 +1,7 @@
 import { DiffSet } from 'dir-compare';
-import { mapDiffSetToPrModule } from './diffset-mapper';
-import { DiffVisitorFn } from './visitors/types';
+import { mapDiffSetToPrModule } from '../../src/diff/diffset-mapper';
+import { DiffVisitorFn } from '../../src/diff/types';
+
 
 describe('Testing the diffset module mapper', () => {
   it('should return an empty array if no visitors are provided', () => {
