+++ packages/amplication-git-pull-request-service/tests/diff/mocks.ts
@@ -1,5 +1,8 @@
 import { Difference } from 'dir-compare';
 import { mock } from 'jest-mock-extended';
+import { BuildPathFactory } from '../../src/diff/build-path-factory';
+
+export const MOCK_BUILD_PATH_FACTORY = mock<BuildPathFactory>();
 
 export const missingFolderMock = mock<Difference>({
   name1: 'name1',
